+++ src/react/components/editabletable/XUIEditableTableFootAction.tsx
@@ -1,43 +1,31 @@
-import cn from 'classnames';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 
 import XUIButton from '../../button';
-import { tableVariantClassNames } from './private/constants';
-import XUIEditableTableUtilityBar from './XUIEditableTableUtilityBar';
 
-const baseName = `${tableVariantClassNames.editable}foot--action`;
-
-interface BaseProps {
-  addButtonContent: string;
+interface Props {
+  buttonContent: string;
   buttonProps?: React.ComponentProps<typeof XUIButton>;
   className?: string;
-  onAdd?: React.MouseEventHandler<HTMLElement>;
+  onClick?: React.MouseEventHandler<HTMLElement>;
   qaHook?: string;
 }
 
-type Props = BaseProps &
-  React.DetailedHTMLProps<React.HTMLAttributes<HTMLTableRowElement>, HTMLTableRowElement>;
-
 const XUIEditableTableFootAction: React.FunctionComponent<Props> = ({
-  addButtonContent,
+  buttonContent,
   buttonProps,
-  className,
-  onAdd,
+  onClick,
   qaHook,
-  ...spreadProps
 }) => (
-  <XUIEditableTableUtilityBar className={cn(baseName, className)} qaHook={qaHook} {...spreadProps}>
-    <XUIButton
-      {...buttonProps}
-      onClick={onAdd}
-      qaHook={qaHook && `${qaHook}--addbutton`}
-      size="small"
-      variant="borderless-main"
-    >
-      {addButtonContent}
-    </XUIButton>
-  </XUIEditableTableUtilityBar>
+  <XUIButton
+    {...buttonProps}
+    onClick={onClick}
+    qaHook={qaHook}
+    size="small"
+    variant="borderless-main"
+  >
+    {buttonContent}
+  </XUIButton>
 );
 
 XUIEditableTableFootAction.propTypes = {
@@ -46,16 +34,15 @@ XUIEditableTableFootAction.propTypes = {
    * <br />
    * Recommended English value: *Add new row*
    */
-  addButtonContent: PropTypes.string.isRequired,
+  buttonContent: PropTypes.string.isRequired,
   /**
    * Props for the row-adding button
    */
   buttonProps: PropTypes.object,
-  className: PropTypes.string,
   /**
    * Function to be fired when the row-adding button is clicked
    */
-  onAdd: PropTypes.func,
+  onClick: PropTypes.func,
   qaHook: PropTypes.string,
 };