+++ src/react/components/editabletable/__tests__/XUIEditableTableFootAction-test.js
@@ -11,43 +11,26 @@ expect.extend(toHaveNoViolations);
 
 describe('<XUIEditableTableFootAction />', () => {
   it('renders correctly', () => {
-    const wrapper = shallow(<XUIEditableTableFootAction addButtonContent="test button" />);
-    expect(toJson(wrapper)).toMatchSnapshot();
-  });
-
-  it('composes the className correctly', () => {
-    const wrapper = shallow(
-      <XUIEditableTableFootAction addButtonContent="test button" className="test-classname" />,
-    );
+    const wrapper = shallow(<XUIEditableTableFootAction buttonContent="test button" />);
     expect(toJson(wrapper)).toMatchSnapshot();
   });
 
   it('should call the onAdd callback when the add button is clicked', () => {
-    const onAddMock = jest.fn();
+    const onClickMock = jest.fn();
     const wrapper = mount(
-      <table>
-        <tfoot>
-          <XUIEditableTableFootAction
-            addButtonContent="test button"
-            qaHook="test-footaction"
-            onAdd={onAddMock}
-          />
-        </tfoot>
-      </table>,
+      <XUIEditableTableFootAction
+        buttonContent="test button"
+        qaHook="test-footaction"
+        onClick={onClickMock}
+      />,
     );
 
-    wrapper.find('[data-automationid="test-footaction--addbutton"]').simulate('click');
-    expect(onAddMock).toBeCalled();
+    wrapper.find('[data-automationid="test-footaction"]').simulate('click');
+    expect(onClickMock).toBeCalled();
   });
 
   it('should pass accessibility testing', async () => {
-    const wrapper = mount(
-      <table>
-        <tfoot>
-          <XUIEditableTableFootAction addButtonContent="test button" />
-        </tfoot>
-      </table>,
-    );
+    const wrapper = mount(<XUIEditableTableFootAction buttonContent="test button" />);
     const results = await axe(wrapper.html());
     expect(results).toHaveNoViolations();
   });