+++ changes.md
@@ -40,6 +40,18 @@ We recommend running a bundle analyser after upgrading (and regularly in general
 
 - `XUICheckboxGroup`, `XUIRadioGroup`, and `XUISwitchGroup` now support horizontal grouping and the `swapAtBreakpoint` to dictate when the group switches from horizontal to vertical grouping. Default behaviour remains unchanged from prior versions.
 
+- `XUIEditableTableFootAction` has been split into two components: `XUIEditableTableFootActions` (the table row) and `XUIEditableTableFootAction` (the action button). This enables multiple action components to be placed inside an editable table footer, where previously only one action was possible.
+  - If you have been using `XUIEditableTableFootAction`, you will need to update your React code. The correct new structure is:
+    ```
+    <XUIEditableTableFoot>
+      <XUIEditableTableFootActions>
+        <XUIEditableTableFootAction .../>
+      </XUIEditableTableFootActions>
+    </XUIEditableTableFoot>
+    ```
+  - If you were providing `className` and `qaHook` props to `XUIEditableTableFootAction`, these props should now be provided to `XUIEditableTableFootActions` instead.
+    - `XUIEditableTableFootAction` now has its own separate `className` and `qaHook` props
+
 ### New components
 
 - `XUIControlGroup` has been introduced. It allows visual and accessible grouping of `XUITextInput`, `XUIDateInput`, `XUIAutocompleter`, `XUISelectBox`, `XUIButton`, `XUIButtonGroup`, and `XUISplitButton`. Additionally, `XUIControlGroup` can be used to group `XUICheckbox`, `XUIRadio`, and `XUISwitch` accessibly but _without_ the visual styling added when using their respective `*Group` components. This replaces the prior CSS-only feature associated with `xui-inputgroup` and `xui-textinputgroup`.
@@ -62,7 +74,10 @@ The following components will ALWAYS have this behaviour from XUI `19.2.0` onwar
 _Note. The codemod will resolve most prop differences automatically when run._
 
 - `XUIDateInput`
-  - Now han an `inputIcon` prop to allow the icon displayed next to the input to be changed
+  - Now has an `inputIcon` prop to allow the icon displayed next to the input to be changed
+- `XUIEditableTableFootAction`
+  - `addButtonContent` has been renamed to `buttonContent`
+  - `onAdd` has been renamed to `onClick`
 
 ## Other changes